Further isolate tmf's core from the rest
No pun intended.

This commit is the starting point to carve tmf's core out of the
current monolithic piece of software.

It probably won't be piece of cake to deal with the fallback since
there are places in the code that I've not deleted that still reference
pieces that I have deleted.

A cursory inspection by @giuseppefutia, in any case, confirms that I
am on track and that I'm keeping only what needs to be kept.