Classify.java: honour Lucene's IMAGE field
Make sure we return the content of the IMAGE field rather than
an empty string to the caller.

Unlike the previous implementation of TMF, however, here we
just return the URL in Lucene rather than the result of scraping
the corresponding webpage on the mobile wikipedia.

The latter step, in fact, is to be performed by the enricher.

While there, sort and reorganize imports.