Call classifyCore() function independently from classify()
With this change, developers can add a personal classification policy between the result of the classification process of a text (in other words the documents obtained through a Lucene query in the CONTEXT field) and the effective output from classifyCore() that retrieve additional information to the URI of the document.

For example, if I want to combine the classification results of chapters of an ebook, having the same output of a traditional classification process, I need to implement some code after calling classify() and before calling classifyCore().